requirements.txt
streamlit==1.31.0
transformers==4.44.0
langchain-groq==0.2.0
langgraph==0.2.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
beautifulsoup4==4.12.3
nltk==3.8.1
sentence-transformers==2.2.2
requests-cache==1.2.0
tf-keras==2.15.0
torch==2.2.2